changing name on createTrip
